==> Step-08-ProcessReport.Rmd <==
---
title: "Research Note"
output:
  html_document:
    theme: united
    df_print: paged
    highlight: tango
    smart: false
    toc: yes
    toc_float: yes
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning=F, message=F, fig.width=8)
```

# Introduction

In this Report we will cover:

(1) Description of dataset
(2) Process of developing the dataset
(3) Summary of data gaps, consistency tests and disambiguation 
(4) Ideas on further questions and potential enhancements

Loading packages and files
```{r}
#setting up the environment
library( dplyr )
library( tidyr )
library( pander )

#update the path with your working directory:
wd <- "/Users/icps86/Dropbox/R Projects/Open_data_ignacio"
setwd(wd)
dir()
```

# 1. Description of dataset 

Sources, key variables, dimensions, etc.
Nonprofit Dataset

For more information on the variables see the **Data Dictionary** [PENDING]


The datasets have been manipulated to remove any sensitive information and to include the following enhancements: 

* Nonprofit mission and taxonomy classification 
* Geographic data (geocodes) for nonprofit addreses and organization member addresses (latitudes and longitudes.
* Census tract corresponding to geocodes and key demographic census data from the corresponding nonprofit and organization member addresses. 
* Voting district correspondnig to geocodes and historic voting results of the district. 
* Gender of organization members (infered from historic us name data).

**Main Characteristics**

a) IDs are repeated so we used input_address as key for the census and google geocoding. 

To make geocoding more efficient, we used input_address as the key. 

* Some cases with duplicate IDs that have the same address written in a slightly different way. Might be useful to understand why the process allows two addresses to be filled - it could be the same business registering twice?





## 1.2  Data Sources

The Nonprofit datasets (Nonprofit and Board Members) have been generated based on IRS' 1023-EZ forms for year 2014 to 2019, which are publicly available [here](https://www.irs.gov/charities-non-profits/exempt-organizations-form-1023ez-approvals).

Nonprofits submit 1023-EZ forms to apply for recognition as a tax-exempt organization under Section 501(c)(3) of the Internal Revenue Code, for more deatils visit the [IRS website](https://www.irs.gov/forms-pubs/about-form-1023-ez). 

Only organizations that meet certain characteristics are elegible, some of the main requierements are:

* Annual gross receipts for the past 3 years and expected gross receipts for thre next 3 years does not exceed $50k
* Assets less than $250,000
* Mailing address inside US territory

For more detail see p. 13 of the [1023-EZ application instructions](https://www.irs.gov/pub/irs-pdf/i1023ez.pdf) 

According to the 1023-EZ instructions, nonprofits must submit detailed organizaiton information, including: the organization's name, mailing address, employee identification number (EIN), mission, contact person, and name, title and mailing address information of up to five organization officers, directors, and/or trustees. 1023-EZ instructions prioritize reporting of organization members in the following order: 

1. President or chief executive officer or chief operating officer.
2. Treasurer or chief financial officer.
3. Chairperson of the governing body.
4. Any officers, directors, and trustees who are substantial contributors (not already listed above).
5. Any other officers, directors, and trustees who are related to a substantial contributor (not already listed above).
6. Voting members of the governing body (not already listed above).
7. Officers (not already listed above).

If an individual serves in more than one office (for example, as both an officer and director), list this individual on only one line and list all offices held.

An officer is a person elected or appointed to manage the organization’s daily operations, such as president, vice president, secretary, treasurer, and, in some cases, board chair. The officers of an organization are determined by reference to its organizing document, bylaws, or resolutions of its governing body, or otherwise designated consistent with state law. A director or trustee is a member of the organization’s governing body, but only if the member has voting rights.


## 1.3  Dataset Representativeness and Relevance

Do the e-filers represent all or most of 1023-EZ filers? What proportion of 1023 filers are EZ filers? What proportion of small nonprofits are recent 1023-EZ filers? Are EZ filers representative of all 1023 filers? If not, what is particular about this sample?

[Pending]

## 1.4 Descriptive Statistics

Loading the data
```{r}
npo1 <- readRDS("Data/3_GeoCensus/NONPROFITS-2014-2019v3.rds")
npo2 <- readRDS("Data/4_GeoGoogle/NONPROFITS-2014-2019v4.rds")
npo3 <- readRDS("Data/5_ZipandCity/NONPROFITS-2014-2019v5.rds")
npo4 <- readRDS("Data/6_CensusData/NONPROFITS-2014-2019v6.rds")

names(npo1)
```

```{r}
ppl1 <- readRDS("Data/3_GeoCensus/PEOPLE-2014-2019v3.rds")
ppl2 <- readRDS("Data/4_GeoGoogle/PEOPLE-2014-2019v4.rds")
ppl3 <- readRDS("Data/5_ZipandCity/PEOPLE-2014-2019v5.rds")
ppl4 <- readRDS("Data/6_CensusData/PEOPLE-2014-2019v6.rds")

names(ppl1)
```


### a) Nonprofit Dataset

* Dimensions
* Nonprofit attributes (taxonomy)
* Geographic data: results of geocode_type and trends in geographic distribution
* Demographic data
* Voting data

Loading the latest npo dataset
```{r}
npo <- npo4
```

#### NPO - Geocoding Types
```{r}
dim(npo)
x <- table(npo$geocode_type, useNA = "ifany")
y <- prop.table(x)
summary <- as.data.frame(t(rbind(x,y)))
colnames(summary) <- c("frequency", "percent")
summary[nrow(summary)+1,] <- c(sum(summary$frequency), 1)
summary$percent <- paste0(round(summary$percent*100,1)," %")
rownames(summary)[nrow(summary)] <- "TOTAL"
summary <- summary[c(1,3,4,2,5,6),]
pander(summary)

```

#### NPO - POBs
```{r, eval=FALSE}
x <- table(npo$pob)
y <- prop.table(x)
summary <- as.data.frame(t(rbind(x,y)))
colnames(summary) <- c("frequency", "percent")
summary[nrow(summary)+1,] <- c(sum(summary$frequency), 1)
summary$percent <- paste0(round(summary$percent*100,1)," %")
rownames(summary) <- c("Non-POB", "POB", "TOTAL")
pander(summary)

```

#### NPO - Locations

Removing unwanted geocodes
```{r}
map <- npo
table(map$geocode_type, useNA = "ifany")

# removing pobs
x <- which(map$pob == 1)
map <- map[-x,]

# removing NAs
x <- which(is.na(map$geocode_type))
map <- map[-x,]

# removing zips and city centroids 
x <- which(map$geocode_type %in% c("zip1", "zip2", "city"))
map <- map[-x,]

# results
table(map$geocode_type, useNA = "ifany")
```

Plotting
```{r}
bkp <- map

# plotting
# plot( map$lon, map$lat, pch=19, cex=0.5, col=gray(0.5,0.01))

# creating variable to remove outliers
map$out <- 0

# trimming lats
max(map$lat)
min(map$lat)

x <- which(map$lat <= 10)
map$lat[x]
map$out[x] <- 1

x <- which(map$lat >= 75)
map$lat[x]
map$out[x] <- 1

# trimming longitudes
max(map$lon)
min(map$lon)

x <- which(map$lon <= -170)
map$lon[x]
map$out[x] <- 1

x <- which(map$lon >= -50)
map$lon[x]
map$out[x] <- 1

# trimming
x <- map$out == 0
map <- map[x,]

plot( map$lon, map$lat, pch=19, cex=0.5, col=gray(0.5,0.01))

```

#### NPO - Top 10 states with more NPOs

```{r}
names(npo)
ste <- npo$State

x <- is.na(ste)
ste <- ste[!x]

ste <- as.data.frame(table(ste), stringsAsFactors = FALSE)
ste <- ste[order(ste$Freq, decreasing = T),]
rownames(ste) <- NULL

ste <- ste[1:20,]

barplot(height = ste$Freq, names.arg = ste$ste)

```

#### NPO - Types of NPOs

```{r}

```

### b) People Dataset

Loading the latest npo dataset
```{r}
ppl <- ppl4
ppl <- unique(ppl)
```

#### PPL - Geocoding Types
```{r}
dim(ppl)
x <- table(ppl$geocode_type, useNA = "ifany")
y <- prop.table(x)
summary <- as.data.frame(t(rbind(x,y)))
colnames(summary) <- c("frequency", "percent")
summary[nrow(summary)+1,] <- c(sum(summary$frequency), 1)
summary$percent <- paste0(round(summary$percent*100,1)," %")
rownames(summary)[nrow(summary)] <- "TOTAL"
summary <- summary[c(1,3,4,2,5,6),]
pander(summary)

```

#### PPL - POBs
```{r, eval=FALSE}
x <- table(ppl$pob)
y <- prop.table(x)
summary <- as.data.frame(t(rbind(x,y)))
colnames(summary) <- c("frequency", "percent")
summary[nrow(summary)+1,] <- c(sum(summary$frequency), 1)
summary$percent <- paste0(round(summary$percent*100,1)," %")
rownames(summary) <- c("Non-POB", "POB", "TOTAL")
pander(summary)

```

#### PPL - Locations

Removing unwanted geocodes
```{r}
map <- ppl
table(map$geocode_type, useNA = "ifany")

# removing pobs
x <- which(map$pob == 1)
map <- map[-x,]

# removing NAs
x <- which(is.na(map$geocode_type))
map <- map[-x,]

# removing zips and city centroids 
x <- which(map$geocode_type %in% c("zip1", "zip2", "city"))
map <- map[-x,]

# results
table(map$geocode_type, useNA = "ifany")
```

Plotting
```{r}
bkp <- map

# plotting
# plot( map$lon, map$lat, pch=19, cex=0.5, col=gray(0.5,0.01))

# creating variable to remove outliers
map$out <- 0

# trimming lats
max(map$lat)
min(map$lat)

x <- which(map$lat <= 10)
map$lat[x]
map$out[x] <- 1

x <- which(map$lat >= 75)
map$lat[x]
map$out[x] <- 1

# trimming longitudes
max(map$lon)
min(map$lon)

x <- which(map$lon <= -170)
map$lon[x]
map$out[x] <- 1

x <- which(map$lon >= -50)
map$lon[x]
map$out[x] <- 1

# trimming
x <- map$out == 0
map <- map[x,]

plot( map$lon, map$lat, pch=19, cex=0.5, col=gray(0.5,0.01))

```

#### PPL - Top 10 states with more PPL residencies

```{r}
ste <- ppl$State

x <- is.na(ste)
ste <- ste[!x]

ste <- as.data.frame(table(ste), stringsAsFactors = FALSE)
ste <- ste[order(ste$Freq, decreasing = T),]
rownames(ste) <- NULL

ste <- ste[1:20,]

barplot(height = ste$Freq, names.arg = ste$ste)

```

#### PPL -Types of Members

```{r}
names(ppl)

table(ppl$Title) # too many categories.
```


# 2. Geocoding Process

* Document decisions
* Provide examples
* Identify outstanding issues
* Challenges / Troubleshooting faced (examples)
* how we deal with special cases (examples)


## 2.1 Geocoding strategy

What is our strategy for enhancing the dataset with census data?

Hierarchy of geocode results:
1. Census and Google Geocode of complete addresses (we need to remove the addresses geocoded with NAs - specially google which can geocode a state)
2. Zipcode centroid
3. City centroid
4. State (might have state but its to broad for geographic analysis)
5. Missing
PENDING: Add a table describing the types of addresses and the geocoding plan for each]


## 2.2 Geocoding Results Summary

## 2.2 Census Geocoding



### What we learn from failed addresses/What causes a failed (No_Match or Tie) in the Census Geocode?
Large office building? Suite #s? 
What else cause a tie? 
PO box only
Incomplete address
Bad street address, but city and zip
No zip, only city
Incorrect spelling
Any others???

### Troubleshooting and things to consider when Geocoding through the Census service
* Sometimes Census can block, or simply output a wrong operation
one whole batch being NULL
* IDs must be unique otherwise the geocode process ignores duplicates. This was a problem because we have some issues with the IDs.

Providing one examples of each type of problem and the solution we applied:

* Tie in census - What else cause a tie? 
* Large office building? Suite #s? 
* PO box only
* Incomplete address
* Bad street address, but city and zip
* No zip, only city
* Incorrect spelling

Some answers are general and some are sapecific to the Census geocode (and other will be to the Google), PENDING to place each in its corresponding place.

## 2.3 Management of Special cases
POBs with complete address
Addresses that seem fake




# 3. Further questions and potential enhancements

What bias could the dataset have?
using POBs (zip centroids) as a source of demographic data


1. What are the main assumptions and/or limitations of the dataset (biases)
  * Business address vs. residential
  * Nonprofit address is related to the organization's target population or member composition - or neither?

2. What questions could arise regarding the dataset generation process, accuracy/reliability, representativeness, etc.? 
3. What additional enhancements could be done to the dataset
4. What sort of research questions could it help answer?
